Step 1: Adds tests for "legendary" items
I used a for comprehension to indicate that the quality should remain
unchanged regardless of the number of times `GildedRose.update_quality/1` is called.

For a more robust test, I'd consider writing a property test to ensure
this behavior, but that is something I do not have much experience with.
Implementing such a test would put me over the recommended 2 hrs.