Wraps legacy update code
Rather than changing the legacy update process I opted to wrap it and
preserve it in a private function. This gives us an easy way to apply
a new update function in a piece-meal fashion as we refactor the special
cases.

Once we can be sure that we've handled the update process for the
various item types we can remove this function.