Step 2: Adds protocol for items
Using a protocol allows us to use the type of a struct to dispatch on
for the various cases needed for inventory items without having to rely
on matching magic strings in the names.

Furthermore, the protocol allows any new "class" of item to be added
without changing any existing code so long as the interface is
respected.